cryptohome: Include verifiers in the auth session start log
When investigating another bug it was noticed that we don't actually
tell you what verifiers are available, only factors. This is not really
an issue with persistent vaults but with ephemeral vaults (which don't
have "real" factors) it's kind of important.

Note that this doesn't really impact the API, the StartAuthSession reply
_does_ contain data from the verifiers. It's just the log message that
was only looking at the auth factor map.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome
TEST=manual log check

Change-Id: If0288c8cf38f855a954afce0979a3b31800052e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5601071
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>