shill: dbus: client: Expose NetworkConfig
Add the code to parse the NetworkConfig property on Service. This
property is proposed to replace the current IPConfig D-Bus interface.
See https://crrev.com/c/5549810 for more details.

The NetworkConfig property will be read and parsed when selected
service is changed or property change signal on a selected service is
received.

BUG=b:340974631
TEST=FEATURES="test" emerge-$BOARD shill-dbus-client

Change-Id: I900923b4eaad33dca1752cd4141024ef354e7e7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5602536
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>