crash: Only kernel crash records matter for kernel crashes
We only return false from CollectRamoopsCrash() if the pstore crash
records, the BIOS log, and the hypervisor dump are all empty after
stripping sensitive data. The only thing that matters for ramoops
recorded crashes is if the ramoops record is empty because that's the
crash we're looking at for the signature. There's almost certainly a
BIOS log all the time as well. Reorganize this code so that only the
ramoops record matters to determine if the crash is empty.

BUG=b:330787908
TEST=cros_sdk cros_run_unit_tests --board=${BOARD} --packages crash-reporter

Change-Id: I49472acc33ccd8a9a6fdef6510927cb7f785e6a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5607242
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>