ml: Add missing syscalls for HeatmapModel
The change in crrev/c/5550738 made sysfs available to the HeatmapModel
and therefore libtouchraw is now able to use the report_descriptor
in sysfs instead of using an IOCTL to request the report descriptor
from the kernel.

This path was previously not enabled and therefore there were
syscalls missing in the seccomp policy file. The missing syscalls
have now been added for arm64 and amd64.

BUG=b:345651478
TEST=`cros build-packages --board=$B chromeos-base/ml && \
     cros deploy ${DUT} chromeos-base/ml` and validate model \
     is running on ciri and rex.

Change-Id: I655dc5f72af84d83431ada271e674aed64372fd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5607151
Tested-by: Henry Barnor <hbarnor@chromium.org>
Reviewed-by: Rob Schonberger <robsc@google.com>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Commit-Queue: Henry Barnor <hbarnor@chromium.org>