shill: Add lease_name field at DHCPCDControllerInterface::Options
DHCPCDControllerInterface::Options is the duplicatation of the current
DHCPController::Options. Originally we thought the lease_name field is
not used anymore, so we didn't copy the field. However, the WiFi class
still uses the field.

This CL adds the field back to DHCPCDControllerInterface::Options, and
implements the related logic at LegacyDHCPCDController.

BUG=b:342068333
TEST=FEATURES=test emerge-brya shill

Change-Id: I2182a2cef4d26f7c13fbdf0e704b2125a8d02dfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5610632
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>