shill: Deregister WiFi device from phy in Manager
Make Manager responsible for initiating WiFi device deregistration from
their associated WiFiPhy. Initiate this from within
Manager::DeregisterDevice. This is the model that we intended from the
beginning, and we're able to implement it now because WiFi registration
has be decoupled from the UP/DOWN state of the device.

This is a superior model because it means that phy registration will now
match 1:1 with manager registration, which occurs in response to the
only two events that we care about: the device being deleted in the
kernel, and the interface being claimed.

BUG=b:248054832
TEST=FEATURES=test emerge-brya shill

Change-Id: I8db6ae7c4efca5af84c7c546d6c8b8beb5c9b910
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5623674
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>