vm_tools: vsh: Fix behavior on shutdown
A planned libchrome CL (crrev.com/c/5501021) will change the behavior of
`base::FileDescriptorWatcher` and it won't work with the current vsh
implementation.
This CL fixes the implementation to match the future behavior.
Specifically, this CL changes the SIGCHLD handler to read all the data
in stdout/stderr fds, rather than leaving it to the fds' handler.

BUG=b:344749812
TEST=`android-sh`
TEST=`android-sh -c 'cat /system/etc/fonts.xml'` for 50 times
TEST=vsh worked with termina

Change-Id: I4000180eac36c1ae885273cfdaf7322d252f65ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5613010
Tested-by: Ryuichiro Chiba <chibar@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: Ryuichiro Chiba <chibar@chromium.org>