system_api: vm_tools: Allow GuestOsSharePath for Fusebox paths
We are going to replace VirtualFileProvider with Fusebox, as the latter
provides a more generalized and better virtual file system integration
with ARC.
The replacement requires sharing Fusebox files with ARCVM, which can
easily be done via seneschal using the GuestOsSharePath mechanism in
Chrome.
This CL adds GuestOsSharePath support for Fusebox paths. The actual
usage from Chrome is found in CL:5610638 and CL:5629191.

BUG=b:346431687
TEST=$ seneschal_client --share_path --vm_name=arcvm --owner_id=<hash> \
       --storage_location=fusebox --path=/
TEST=See CL:5610638 and CL:5629191.

Change-Id: I9d2d59c90f8e661abe3b058ac1d229255e0696ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5612955
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>