crash: Add support for efi_pstore driver name
With the linux commit 893c5f1de620fb0134ddef37dfd850f6c639162b, the
efi pstore driver has been renamed from "efi" to "efi_pstore". This
driver name is used as a pattern when collecting and processing kernel
crash files, so we must make both patterns  work in order to be
compatible with both newer and older kernel versions.

Specifically, we search for files matching either dmesg-efi-* or
dmesg-efi_pstore-*. This driver name is saved on the EfiCrash instance
and is used when working with paths for that crash.

The tests have also been updated to try both the efi and efi_pstore
driver names to ensure logs are collected in both cases.

BUG=None
TEST=cros_sdk cros_run_unit_tests --board=${BOARD} --packages
crash-reporter

Change-Id: If04288ac56917bb4ba42c79ab28ff5225fb9b2bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5598051
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>