patchpanel: check client validity during socket annotation
As it is not possible to bind a WeakPtr to a callback with a return
value, we were binding a base::Unretained() reference to the ClientImpl
instance to obtain a valid 'this' reference in
ClientImpl::OnSocketAnnotation.  Since the callback stays registered in
the HTTP transport even if the client is destroyed, it can lead to a
crash when the callback is fired.

To avoid the crash, the OnSocketAnnotation() callback is changed to a
free function and the reference to the ClientImpl is passed as a
parameter, as a WeakPtr. The callback now checks the validity of the ptr
before calling TagSocket().

BUG=b:345769752
TEST=FEATURES=test emerge-$BOARD chromeos-base/patchpanel-client
  Deploy on a DUT and verify using the log that TagSocket() is correctly
  called during Shill's captive portal checks.

Change-Id: I39ceab15cc94bd1bf8ccc1585b464c7c2657012d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5620291
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Damien Dejean <damiendejean@google.com>
Commit-Queue: Damien Dejean <damiendejean@google.com>