diagnostics: Extend timeout for cURL failure
We found some cURL timeout error cases when running network bandwidth
routine in Lab. It usually took about 0.1 second to complete cURL
process and the timeout was 7 seconds.

Since we got a case that took 4 seconds to complete cURL process when
network speed is fast (~300 Mbps), the 7 seconds timeout could be not
enough.

Extend the timeout to 10 seconds to check if the cURL timeout error can
be mitigated.

BUG=b:336179800
TEST=FEATURES=test emerge-volteer diagnostics
TEST=(On DUT) cros-health-tool diag network_bandwidth

Change-Id: Ide626af8f00e0711bc04cb6902f0a5f0612953f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5618214
Reviewed-by: Wei-Luan Wang <weiluanwang@google.com>
Commit-Queue: Byron Lee <byronlee@chromium.org>
Tested-by: Byron Lee <byronlee@chromium.org>