vm_tools: Disable pvclock for x86_64
As pvclock seems to be the culprit for b/342168920, disable it in the
meantime.

This is a revert of f46bbac797e ("vm_tools: concierge: Append --pvclock
crosvm arg on x86 ARCVM") but in a separate commit since the change
doesn't revert cleanly anymore.

BUG=b:295256641,b:346249019
TEST=tast run arc.Suspend.s10c2

Cq-Depend: 5624354
Change-Id: I57e7a13445025a2fe9021fca956aa97663ee1dfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5622438
Commit-Queue: Tatsuyuki Ishi <ishitatsuyuki@google.com>
Tested-by: Tatsuyuki Ishi <ishitatsuyuki@google.com>
Reviewed-by: Suleiman Souhlal <suleiman@google.com>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>