fpstudy: Rename TestCase class to TestCaseEnum
We are not going to follow this pattern of statically defined
test cases in code. Instead, the TestCase class will be recycled
to represent the name, description, and *associated Experiment*
of a real test case.

I initially tried to add a common interface for TestCaseEnum and a
new generic TestCaseDescriptor class, but it was too cumbersome
to make an TestCaseEnum a subclass of both Enum and an interface for
TestCaseDescriptor. For one, it becomes difficult to provide a name()
accessor, since Enum already defines a ".name" @property. Then, I ran
into issues making a reasonable constructor. This inheritance/hierarchy
felt forced just to allows coded Enums, when I really don't want future
users to follow this in-code Enum TestCase pattern, anyways.

BUG=b:339298572
TEST=python -m unittest discover -v -s . -p '*_test.py'
TEST=./run.py report --user-groups-csv="${test_cases}/user_groups.csv" \
       "${test_cases}" "${output}"

Change-Id: I318f24df478af6781aa0ea045a3ec8b95d023957
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5622601
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Firas Sammoura <fsammoura@google.com>
Commit-Queue: Craig Hesling <hesling@chromium.org>