installer: Sync before calling postinstall
Without this change, postinstalls fail to open the kernel partition
device on internal storage. There was a proper fix for this issue in
crrev/c/5587971, which did break other parts of chromeos-install though.

According to the discussion in
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5496291/comment/588a8369_7be3d630/
it's also a solution to call `sync` before starting postinstalls, as
this presents enough of a delay. This is not a nice fix for the
underlying issue, but rather a temporary change to get chromeos-install
fixed for all its usecases.

BUG=b:346730882
TEST=manual

Change-Id: I41af5961c5f04635ce1ef91c9540a16874a462ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5621207
Tested-by: Bastian Kersting <bkersting@google.com>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Bastian Kersting <bkersting@google.com>
Reviewed-by: Thiago Bastos <tbastos@google.com>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Auto-Submit: Bastian Kersting <bkersting@google.com>