cryptohome: Add Ready functions  to PrepareAuthFactorToken
This CL lays the foundation for using a state in PrepareAuthFactor for
SmartCard.

The two states, IsReadyForClient will signify that the token can be
sent back to the client.

The IsReadyForNextAuthOperation  will mean that the token can be
is ready to be used in the next step, which can be authenticate,
add, updates, etc.

BUG=b:270631179
TEST=FEATURES=test emerge-zork cryptohome

Change-Id: I4b693dbb08e3bc0a7a9b80d59ea4a58e4ebafdd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5633425
Auto-Submit: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: John Admanski <jadmanski@chromium.org>