shill: Fix race condition in P2PManager timeout flow
Currently, the P2PManager action timer is not armed until after the
P2P device creation request has been sent to the manager. This request
triggers an asynchronous attempt to create the interface. If this
request completes prior to the timer being armed, we can end up in a
situation where the timer gets armed after the request is complete, and
thus fires unexpectedly in the future.

Arm the action timer before requesting the device creation to prevent
this. This means that we also have to disarm the action timer if our
request is rejected.

Also fixes a minor typo in P2PManager where the P2PGO device type was
being used for arming a P2PClient timer. This typo had no functional
impact because the underlying D-bus reason codes resolve to the same
strings for both types, but its still good to fix.

BUG=b:323064949
TEST=FEATURES=test emerge-brya shill

Change-Id: I1f4f97facd3cb00426dc2334d3502086a638c952
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5630696
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>