dns-proxy: Treat NXDOMAINs as invalid response
When multiple nameservers are set, DNS proxy queries the nameservers in
parallel. This can cause issue if the target domain is only resolvable
by some nameservers. When this happen, if "bad" nameserver returns a
response before the "good" nameserver with NXDOMAIN, DNS proxy will use
the NXDOMAIN answer even if there is a better answer.

This patch updates DNS proxy's behavior such that NXDOMAINs are seen as
error and would wait for a better answer if possible.

BUG=b:333757554
BUG=b:345555156
TEST=manual; see that NXDOMAINs are not returned directly.

Change-Id: I882022dc2c210a23cd359b6596fb12af39b894a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5642118
Commit-Queue: Jason Iman <jasongustaman@chromium.org>
Tested-by: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>