Remove nverne@chromium.org from vm_tools/OWNERS
This suggested change is automatically generated based on commit
and review activity to improve the reliability of OWNERS suggestions
and CL review latency.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.
If no CR -1 vote is set after 5 days, the CL will be submitted.
OWNERS expectations for the chromium project are at:
https://chromium.googlesource.com/chromium/src/+/refs/heads/main/docs/code_reviews.md#expectations-of-owners

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/nverne@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I39109865d02c2657fae8059c3e6fcd2a6404b82f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5647171
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Tested-by: Nicholas Verne <nverne@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Nicholas Verne <nverne@chromium.org>