shill: dbus: client: Refactor HandleSelectedServiceChanged()
Avoid doing another lookup of DeviceWrapper inside this function. The
current two callers already has the DeviceWrapper object so pass it to
this function directly.

Move the DeviceWrapper definition to the beginning of the private
section of Client so that it can be used in the declaration of this
function.

BUG=b:348575562
TEST=FEATURES="test" emerge-$BOARD shill-dbus-client

Change-Id: I1eb9eb62adb6f5e6531ad7342e6e2072f3c1cdf0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5645274
Reviewed-by: Jason Iman <jasongustaman@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>