cryptohome: Remove the restore key intent and RPC
We're not going ahead with this design, so lets start pruning out the
implementation. This removes the intent & the restore key, I'll take a
look at evict key in a followup.

This doesn't remove the proto definitions because there's some chromium
references to them that will need to be deleted first.

BUG=None
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Icd1b3b786266e06a0144cb320b937f86304fd85a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5649010
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>