adbd: Disable DbC on USB-C bus when no typec cables present
When DbC is enabled on a USB bus, it polls continually, looking for
new connections. Currently DbC is only used for ADB debugging on
ChromeOS. This patch disables DbC when ADB debugging when no
typec cables are connected to prevent needless polling and energy
consumption.

When the first cable is attached, enable DbC.

To determine when the first cable is attached / last cable is
removed, this uses a simple counting mechanism, looking for
portX-partner events on the bus configured in /etc/arc/adbd.json.

BUG=b:121287022
TEST=Check the status of the DbC entry when ADB is en/disabled:
  No cable (including typec power cables):
  cat /sys/bus/pci/devices/$DEVICE_ID/dbc
  disabled

  Plug in cable:
  cat /sys/bus/pci/devices/$DEVICE_ID/dbc
  enabled
  (or configured)

  Remove cable:
  cat /sys/bus/pci/devices/$DEVICE_ID/dbc
  disabled

  $DEVICE_ID is the USB bus id. For brya/volteer: 0000:00:0d.0

Change-Id: I301e3f7f4f74f08c47d80e0179c85e879e4cf18b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5654290
Reviewed-by: Emilie Roberts <hadrosaur@google.com>
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Commit-Queue: Emilie Roberts <hadrosaur@google.com>
Tested-by: Emilie Roberts <hadrosaur@google.com>