camera: algo: use _exit() in when exit the camera algo process for libcamera
Ciri libcamera only starts the camera algo process when the camera is
opened to avoid extra memory usage, and close the process when the
camera is closed. It's noticed that the close sourced algo library would
crash sometimes when the process calls exit(). The patch replace exit()
with _exit() to prevent calling atexit from the closed library to
avoid the issue.

BUG=b:349555857
TEST=CTS CtsCameraTestCases and no crash reports for algo

Change-Id: I14f390c36d43516e4c1de69ad78e7b510bfddfdb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5656904
Commit-Queue: Hanlin Chen <hanlinchen@google.com>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Reviewed-by: Harvey Yang <chenghaoyang@chromium.org>
Tested-by: Hanlin Chen <hanlinchen@google.com>