shill: wifi: skip STA/AP concurrency check on tethering downstream technology
Previously, AP mode needs a primary STA interface coexist in order to
start. With the concurrency logic, the STA interface will be destroyed
to make room for the SoftAP interface if concurrency is not supported.
Thus there is no need to check STA/AP concurrency (which always return
false as all the boards do not support DBS) when deciding the supported
downstream technologies.

BUG=b:349678312
TEST=on brya, check if hotspot entry is available in the setting page

Change-Id: Ib53d77dfabec293f2e1cfe1c272fc757e2474cd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5659969
Tested-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Kevin Lund <kglund@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Jintao Lin <jintaolin@chromium.org>