rmad: Make an empty SKU filter have no effect.
An empty SKU filter could be a placeholder for partners to fill in the
correct SKUs in the future. Since it doesn't make sense to create a SKU
filter that actually filters out all SKU, we make the effect of an empty
SKU to have no effect, rather than showing nothing in the SKU list.

BUG=b:349412977
TEST=FEATURES=test emerge-dedede rmad

Change-Id: Ib6eda369183ae87c488007434db59e5e2ded5292
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5665343
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>
Reviewed-by: Jeff Lin <jeffulin@google.com>
Tested-by: Cheng-Han Yang <chenghan@chromium.org>