shill: Add ability to cancel pending device requests in WiFiProvider
Adds a mechanism for cancelling pending device creation requests in
WiFiProvider. Exposes a public function in WiFiProvider that can be used
to cancel all pending requests of a given type. This is necessary for
two reasons: 1. So that we can cancel requests when they time out. 2. So
that we can cancel pending STA requests when the priority of an
interface changes.

This is - by design - a suboptimal approach. Ideally, we'd be able
cancel only a specific request, rather than always cancelling all
requests of a given type. However, such an implementation would be
significantly more complex, and wouldn't actually provide any benefit
for our existing use-cases. All existing local interface creation callers
allow only a single request at a time, and we only support a single STA
interface.

This solution strongly favors the side of safety and simplicity. The ROI
of building a more robust solution is simply not worthwhile, in my
opinion, particularly in light of recent priority shifts.

BUG=b:328074585
TEST=FEATURES=test emerge-brya shill

Change-Id: I874b90cde63f8597ef06c686c2c2ca639fa29674
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5673592
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Andrzej Ostruszka <andrzejo@google.com>
Tested-by: Kevin Lund <kglund@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>