ethernet-hide: Notify users of the disconnecting SSH
This patch notifies users of the disconnecting SSH before ehide starts
or stops to avoid surprising them. Since the ehide daemon cannot print
anything to the terminal after daemonization, we need to print the
notification before daemon setup and teardown. To achieve this, we
created pre_set_up() and pre_tear_down() in the Daemon class and
implemented them in the EhideDaemon class. They will be called right
before the daemon setup and teardown.

BUG=b:258359554
TEST=observe the notification in terminal
TEST=network.EhideNoEthernet

Change-Id: Ie384ebccc8b6aa8a639f2d3c77ac87303cfcca83
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5677123
Commit-Queue: Zikai Chen <chenzikai@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Zikai Chen <chenzikai@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>