ethernet-hide: Avoid init check when executing "ehide state"
Currently ehide performs init check as long as current state is "off".
This means that ehide will perform this check even when the user only
wants to check its state by executing "ehide state", as long as it is
not running. This is meaningless and will produce a lot of garbage in
the log file.

This patch fixes this by making ehide only do init check when the action
is "start".

BUG=b:258359554
TEST=confirm manually that no init check is performed for "ehide state"
TEST=tast run ${DUT} network.EhideNoEthernet

Change-Id: I74a56dff54db6ecd5448d051529deafac6fc0454
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5677692
Commit-Queue: Zikai Chen <chenzikai@google.com>
Tested-by: Zikai Chen <chenzikai@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>