biod: updater: Don't use CrosConfig in FindFirmwareFile function
The function is responsible for finding appropriate firmware and should
take board name as an input. Using CrosConfig in this function only
complicates this function and violates Single Responsibility Principle.
Moreover it prints information about board which will be weird if the
function is used in other places.

BUG=b:347224521
TEST=FEATURES=test emerge-zork biod

Change-Id: I9f33be510e3e63c25287b96766054ceab13cb9d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5684612
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Tested-by: Patryk Duda <patrykd@google.com>
Commit-Queue: Patryk Duda <patrykd@google.com>