odml: Add support for new on-device model API
Port from crrev.com/c/5675806

The main difference here is that the new internal API is based on a
"session" object. The internal session methods can be blocking, so have
to be run off the main thread.

This keeps things in a state where it will still work with the old API,
and will switch over to the new API once it is available in the internal
lib.

BUG=b:350812364
TEST=Run the odmld & odml_console, and check the result.

Change-Id: I2054d29898d1cb19d15318786760a1064fbf8e1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5679984
Commit-Queue: Yi Chou <yich@google.com>
Reviewed-by: Jim Pollock <jmpollock@chromium.org>
Tested-by: Yi Chou <yich@google.com>