cryptohome: Delete rpc.proto copy of CryptohomeErrorCode
This copy of the error codes is obsolete, only the one in the
user_data_auth package should be used. This does require deleting a few
stray references to the rpc.proto versions from a few unit tests.

This bundles together the API change and the cryptohome change because
the API changes are heavily coupled to the static asserts in cryptohome
and it doesn't really make sense to separate the removal.

BUG=b:135984863
TEST=built all cryptohome code
TEST=built chromium code with changes applied

Change-Id: I65de67412cc5fafbae1789e37fd835e8ba4baa2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5690410
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>