chromeos-dbus-bindings: Use proto includes in adaptor headers as well
This plugs the support added to proxy.go into adaptor.go as well, since
it needs the proto includes for the same reason. The helper added for
this is also moved into genutil, so that it can be more easily reused by
both the adaptor and proxy subpackages.

While fixing this up this also applies the same change to proxy/mock.go
and it also tweaks the includes by sorting them as well.

BUG=b:349775903
TEST=unit tests, CQ

Change-Id: Ic35d8ceab375e9b5df563ec36acbbeaa736903a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5692017
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>