system_api, vm_tools: Add pvclock feature flag to ARCVM request
This enables virtio-pvclock to be controlled through a Chrome feature
flag, for controlled rollout and killswitching.

The macro USE_VIRTIO_PVCLOCK is renamed to ALLOW_VIRTIO_PVCLOCK and now
acts as a gate for the feature. If ALLOW_VIRTIO_PVCLOCK is set to 0,
pvclock is disabled regardless of feature flag specified by Chrome.

BUG=b:342168920
TEST=CQ

Change-Id: I13be6416f6cfe8019d99f8fbab034cc49684d8f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5692241
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Tested-by: Tatsuyuki Ishi <ishitatsuyuki@google.com>
Commit-Queue: Tatsuyuki Ishi <ishitatsuyuki@google.com>