shill: Check state in Network::SetupConnection()
Call this function in the idle state is unexpected, which will
eventually cause a crash in OnSetupConnectionFinished().

Note that OnSetupConnectionFinished() is bound with
weak_factory_for_connection_, and Network::StopInternal() will cancel
all weak ptrs created by weak_factory_for_connection_, so as long as we
add check in SetupConnection(), it should be safe on this path.

BUG=b:352316121
TEST=CQ

Change-Id: Ib77c2cead0c61058939aae37fbaa68930741ca8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5693362
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>