shill: device: create implicit Network in subclasses
This patch moves the creation of the implicit primary network of a
Device to the constructors of the child classes. In addition, the
interface index and interface name are passed explicitly as arguments.
This allows the Device to not have to remember about the interface name
and interface index for the purpose of creating its implicit Network.

BUG=b:352665085
TEST=FEATURES=test emerge-$BOARD shill, CQ

Change-Id: If58c100af0a7934c845458b05ff310dd28bc7cab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5711236
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>