Modify all DBus interface ProtobufClass annotations to include headers.
This change is necessary for the header produced by the code generator
to actually be robust and not depending in include order. More
specifically, this avoids the problem where the generated headers will
fail to compile if the protobuf headers they depend on are not included
first.

Doing a build with these changes should be sufficient to verify that
there are no syntactical problems with this change, and that all of the
reference headers exist. Unfortunately, there's no large scale way to
verify that these are the _correct_ headers for each message, this is
relying on the best efforts of the CL author.

BUG=b:349675903
TEST=CQ

Change-Id: I8796afc8ea201009806413bbf6f0ffb2097c9e3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5703957
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Reviewed-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Andrew Bregger <andrewbregger@google.com>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>
Reviewed-by: Chris Mullins <crmullins@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>