shill: device: separate "name" and interface name
This patch separates inside the base Device class the "Name" property
from the "Interface" property with two separate variables. The
constructor and comments are updated in the follow-up patch that
removes the link_name_ variable from the base Device class. Logs are
updated to always be consistent with the Name property by using
LoggingTag().

This separation allows a Cellular Device to register a name at
construction time that may be unrelated to any network interface. It
also removes most of the usage of link_name_ inside Device.

BUG=b:352665085
TEST=FEATURES=test emerge-$BOARD shill, CQ

Change-Id: If0142ebab00c2b0a5262d924cb7be88ba8f4e4f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5711237
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>