shill: metrics: Add a metric to log AP 6GHz support
Adds a metric for 6GHz band support. Report that an AP
supports 6GHz if:
- the DUT is connected to the AP on the 6GHz band
- otherwise, check RNR of the connected AP to see
if it is colocated with a 6GHz AP.

BUG=b:319927708
TEST=emerge and deploy shill changes to DUT
Then, connected DUT to an AP on 6GHz/ a 5GHz AP which was
colocated with a 6GHz AP
Confirmed that "NotifyBand6GHzSupport: 1" was present in
net.log.
Also, applied chromium:5710173 and checked the local
histogram on the DUT at chrome://histograms.

Change-Id: I6b07cb77bb50ac0290a934d1b6f5ac09acd29de3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5710248
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Ruth Mekonnen <rmekonnen@google.com>
Tested-by: Ruth Mekonnen <rmekonnen@google.com>