shill: Enable RFC8925 on WiFi service conditionally
Currently CrOS does not support IPv6-only network with a link-local DNS
server, since dnsproxy cannot contact with that server, and thus the
network will not be usable if it's the case (see $BUG for details). This
issue becomes more common after we turned on RFC8925 by default since it
may drop IPv4 if the network claims it supports that.

This patch provides a workaround for this scenario by disabling RFC8925
by default for each WiFi service, and turning it on in the second
connection if we found that we can run IPv6-only mode on this network
during the first connection. See inline comments for the detailed logic.

We will be able to remove this workaround after dnsproxy can handle IPv6
link-local name servers properly.

BUG=b:345372970
TEST=FEATRURES="test" emerge-$BOARD shill
TEST=Connected to the WiFi for the second time, saw that RFC8925 is
  enabled. Verify that it's still enabled on the first connection after
  reboot.
TEST=New tast test added in https://crrev.com/c/5717670

Change-Id: I86d3c2f15fdbe256eeed41756bb7d8b24649aa3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5711233
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>