patchpanel: initialize services in member initialized list
Originally, Manager initializes and destroys the service instances
manually inside constructor and destructor. It may cause the
null-pointer dereference crash if the initialization order is wrong.

This CL changes to initialize the data members at member initialized
list. Then the compiler checks the initialization order automatically.
For example, if B depends on A, then A should be initialized earlier
then B.

Also, this CL changes to destroy the data members implicitly. Then the
instances are destroyed at the reversed order of initialization. It
guarantees A's lifetime is longer than B if B depends on A.

BUG=b:354063244
TEST=FEATURES=test emerge-nissa patchpanel

Change-Id: Ib23f0681c83eaf53638625a667f4e4bc1a02bc23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5715591
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>