dns-proxy: Avoid referencing an invalid pointer
The data owned by `flag_fd` is referenced in `argv` which will be used
outside the if branch, but `flag_fd` won't outlive the if branch, and
thus it's actually a dangling pointer. Move the definition of `flag_fd`
outside the branch to avoid this problem.

BUG=None
TEST=FEATURES="test" emerge-$BOARD chromeos-base/dns-proxy

Change-Id: I7e2755c7b0ea24465a2ec5386f1dda240edd56a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5720270
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>