dns-proxy: Add support for enabling VLOG in upstart service
This patch adds a cmdline flag of `--vmodule` in the upstart service
config file for dnsproxy, so that dnsproxy can be restarted with
enabling VLOG for debugging. Also pass this flag from the controller
process to the children processes, since it won't be inherited by
default.

There is no default behavior change.

BUG=b:351994513
TEST=Manually tested with the following operations:
  - Reboot DUT, observed no verbose logs in net.log;
  - `restart dns-proxy VMODULE_ARG=*=2`, observed verbose logs in
    net.log;
  - Checked the `ps` result for `dnsproxy`, all processes have the
    expected `--vmodule` flags

Change-Id: If8e8de328ddf3f747c01adb37e82770661fed418
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5717919
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Jason Iman <jasongustaman@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>