cryptohome: Drop the AuthScanResult signal entirely
This signal is superceded by the PrepareAuthFactorProgress signal. The
fingerprint auth process was already updated to send the new signal in a
previous change, this does the second half of the transition by removing
the old signal.

Note that this does _not_ remove the AuthScanResult proto message
itself. This message is in fact a sub-message of the larger prepare
signal and so it's not expected to go away entirely.

BUG=b:347752239
TEST=CQ, unit tests

Change-Id: If0e87327ed98a1db120845d7b21bb933d44fea48
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5734961
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>