shill: cellular: Add flag to enable/disable the fallback APN
Add a flag in the modb, to allow configuring the use of the fallback
APN on a per carrier basis.
The reason to add this, is that some carriers use OTA APN provisioning,
and using an empty APN can break the provisioning.

BUG=b:344519124
BUG=b:343028329
TEST=Verify that the attach and default APN try list on Verizon doesn't
include the fallback APN, while the change doesn't affect a AT&T DUT.

Change-Id: I31ecac2097a983dbc79439ae326a0aba6a18ecbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5743727
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>